+++ packages/tools/utils/src/modules/any-undefined.js
@@ -1,0 +1,19 @@
+/**
+* @module @svizzle/utils/any-undefined
+*/
+
+/**
+ * A function that does nothing
+ *
+ * @function
+ * @arg {*} any
+ * @return {undefined} - Any -> Undefined
+ *
+ * @example
+> noop()
+undefined
+ *
+ * @version v0.14.0
+ */
+// eslint-disable-next-line no-empty-function
+export const noop = () => {};