+++ packages/tools/utils/src/modules/any-undefined.spec.js
@@ -1,0 +1,15 @@
+import {strict as assert} from 'assert';
+
+import {noop} from './any-undefined';
+
+describe('Any -> Undefined', function() {
+	describe('noop', function() {
+		it('should do nothing and return `undefined`', function() {
+			assert.deepStrictEqual(noop(), undefined);
+			assert.deepStrictEqual(noop(1), undefined);
+			assert.deepStrictEqual(noop('d'), undefined);
+			assert.deepStrictEqual(noop([1, 2]), undefined);
+			assert.deepStrictEqual(noop({a: 1}), undefined);
+		});
+	});
+});