+++ packages/tools/utils/CHANGELOG.md
@@ -1,6 +1,15 @@
-## `@svizzle/utils` v0.13.1 (next)
+## `@svizzle/utils` v0.14.0 (next)
 
 - updated some dev dependencies
+
+### Any -> Undefined
+
+- added `noop`
+
+### Array -> (Any-Any)
+
+- added `truthynessTo`
+
 
 ## `@svizzle/utils` v0.13.0